+++ android/src/main/java/xyz/justsoft/video_thumbnail/VideoThumbnailPlugin.java
@@ -1,5 +1,6 @@
 package xyz.justsoft.video_thumbnail;
 
+import android.content.Context;
 import android.graphics.Bitmap;
 import android.media.MediaMetadataRetriever;
 import android.os.Handler;
@@ -32,11 +33,13 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
     private static String TAG = "ThumbnailPlugin";
     private static final int HIGH_QUALITY_MIN_VAL = 70;
 
+    private Context context;
     private ExecutorService executor;
     private MethodChannel channel;
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+        context = binding.getApplicationContext();
         executor = Executors.newCachedThreadPool();
         channel = new MethodChannel(binding.getBinaryMessenger(), "plugins.justsoft.xyz/video_thumbnail");
         channel.setMethodCallHandler(this);
@@ -73,7 +76,7 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
                 try {
                     if (method.equals("file")) {
                         final String path = (String) args.get("path");
-                        thumbnail = buildThumbnailFile(video, path, format, maxh, maxw, timeMs, quality);
+                        thumbnail = buildThumbnailFile(video, headers, path, format, maxh, maxw, timeMs, quality);
                         handled = true;
 
                     } else if (method.equals("data")) {
@@ -113,10 +116,11 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
         }
     }
 
-    private byte[] buildThumbnailData(String vidPath, HashMap<String, String> header, int format, int maxh, int maxw, int timeMs, int quality) {
-        Log.d(TAG, String.format("buildThumbnailData( format:%d, maxh:%d, maxw:%d, timeMs:%d, quality:%d )", format,
-                maxh, maxw, timeMs, quality));
-        Bitmap bitmap = createVideoThumbnail(vidPath, header, maxh, maxw, timeMs);
+    private byte[] buildThumbnailData(final String vidPath, final HashMap<String, String> headers, int format, int maxh,
+            int maxw, int timeMs, int quality) {
+        // Log.d(TAG, String.format("buildThumbnailData( format:%d, maxh:%d, maxw:%d,
+        // timeMs:%d, quality:%d )", format, maxh, maxw, timeMs, quality));
+        Bitmap bitmap = createVideoThumbnail(vidPath, headers, maxh, maxw, timeMs);
         if (bitmap == null)
             throw new NullPointerException();
 
@@ -128,14 +132,20 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
         return stream.toByteArray();
     }
 
-    private String buildThumbnailFile(String vidPath, String path, int format, int maxh, int maxw, int timeMs,
-                                      int quality) {
-        Log.d(TAG, String.format("buildThumbnailFile( format:%d, maxh:%d, maxw:%d, timeMs:%d, quality:%d )", format,
-                maxh, maxw, timeMs, quality));
-        final byte bytes[] = buildThumbnailData(vidPath, new HashMap<String, String>(), format, maxh, maxw, timeMs, quality);
+    private String buildThumbnailFile(final String vidPath, final HashMap<String, String> headers, String path,
+            int format, int maxh, int maxw, int timeMs,
+            int quality) {
+        // Log.d(TAG, String.format("buildThumbnailFile( format:%d, maxh:%d, maxw:%d,
+        // timeMs:%d, quality:%d )", format, maxh, maxw, timeMs, quality));
+        final byte bytes[] = buildThumbnailData(vidPath, headers, format, maxh, maxw, timeMs, quality);
         final String ext = formatExt(format);
         final int i = vidPath.lastIndexOf(".");
         String fullpath = vidPath.substring(0, i + 1) + ext;
+        final boolean isLocalFile = (vidPath.startsWith("/") || vidPath.startsWith("file://"));
+
+        if (path == null && !isLocalFile) {
+            path = context.getCacheDir().getAbsolutePath();
+        }
 
         if (path != null) {
             if (path.endsWith(ext)) {
@@ -158,7 +168,7 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
             f.close();
             Log.d(TAG, String.format("buildThumbnailFile( written:%d )", bytes.length));
         } catch (java.io.IOException e) {
-            e.getStackTrace();
+            e.printStackTrace();
             throw new RuntimeException(e);
         }
         return fullpath;
@@ -196,7 +206,8 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
      * @param targetH the max height of the thumbnail
      * @param targetW the max width of the thumbnail
      */
-    public static Bitmap createVideoThumbnail(final String video, final HashMap<String, String> header, int targetH, int targetW, int timeMs) {
+    public Bitmap createVideoThumbnail(final String video, final HashMap<String, String> headers, int targetH,
+            int targetW, int timeMs) {
         Bitmap bitmap = null;
         MediaMetadataRetriever retriever = new MediaMetadataRetriever();
         try {
@@ -205,17 +216,14 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
             } else if (video.startsWith("file://")) {
                 setDataSource(video.substring(7), retriever);
             } else {
-                if (header != null) {
-                    retriever.setDataSource(video, header);
-                } else {
-                    retriever.setDataSource(video, new HashMap<String, String>());
-                }
+                retriever.setDataSource(video, (headers != null) ? headers : new HashMap<String, String>());
             }
 
             if (targetH != 0 || targetW != 0) {
                 if (android.os.Build.VERSION.SDK_INT >= 27 && targetH != 0 && targetW != 0) {
                     // API Level 27
-                    bitmap = retriever.getScaledFrameAtTime(timeMs * 1000, MediaMetadataRetriever.OPTION_CLOSEST, targetW, targetH);
+                    bitmap = retriever.getScaledFrameAtTime(timeMs * 1000, MediaMetadataRetriever.OPTION_CLOSEST,
+                            targetW, targetH);
                 } else {
                     bitmap = retriever.getFrameAtTime(timeMs * 1000, MediaMetadataRetriever.OPTION_CLOSEST);
                     if (bitmap != null) {
@@ -251,7 +259,7 @@ public class VideoThumbnailPlugin implements FlutterPlugin, MethodCallHandler {
         return bitmap;
     }
 
-    private static void setDataSource(String video, MediaMetadataRetriever retriever) throws IOException {
+    private static void setDataSource(String video, final MediaMetadataRetriever retriever) throws IOException {
         File videoFile = new File(video);
         FileInputStream inputStream = new FileInputStream(videoFile.getAbsolutePath());
         retriever.setDataSource(inputStream.getFD());